+++ SekiroTool/Services/EnemyService.cs
@@ -9,25 +9,25 @@ namespace SekiroTool.Services;
 public class EnemyService(IMemoryService memoryService, HookManager hookManager) : IEnemyService
 {
     public void ToggleNoDeath(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoDeath, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoDeath, isEnabled ? 1 : 0);
 
     public void ToggleNoDamage(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoDamage, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoDamage, isEnabled ? 1 : 0);
 
     public void ToggleNoHit(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoHit, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoHit, isEnabled ? 1 : 0);
 
     public void ToggleNoAttack(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoAttack, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoAttack, isEnabled ? 1 : 0);
 
     public void ToggleNoMove(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoMove, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoMove, isEnabled ? 1 : 0);
 
     public void ToggleDisableAi(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.DisableAi, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.DisableAi, isEnabled ? 1 : 0);
 
     public void ToggleNoPostureBuildup(bool isEnabled) =>
-        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoPosture, (byte)(isEnabled ? 1 : 0));
+        memoryService.WriteUInt8(DebugFlags.Base + (int)DebugFlags.Flag.AllNoPosture, isEnabled ? 1 : 0);
     
     public void ToggleTargetingView(bool isEnabled)
     {
